requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
PyYAML==6.0
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0